pyproject.toml
[tool.poetry]
name = "limber-timber"
version = "0.1.0"
description = ""
authors = ["Daniel Tashjian <thewopple@gmail.com>"]
readme = "README.md"
packages = [
    { include = "liti" },
]

[tool.poetry.scripts]
liti = "liti.cli:main"

[tool.poetry.dependencies]
python = "3.13.0"
google-cloud-bigquery = "3.34.0"
pydantic = "2.11.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
